Refactored manifest task to gatewat
